Fix maven-compile-plugin version and target. Update code.
After exploring the Java's source code and wondering why on Earth "mvn compile" was giving an error that it cannot cast to Object[],
when it should be compiling just fine, before proceeding to submit a bug complaint regarding the java compiler (joking) I figured that
this had something to do with maven. Well, turns the maven-compile-plugin was not specified as a dependency and the one used by default
was using java 1.3 by default to compile the code, so yeah... Varargs were introduced in 1.5. I didn't speicify a newer version as a
dependency (since it wasn't in the provided pom.xml and I'm not sure if we can add dependencies for the lab), I did however instructed it
to use java 1.8 as target and source. That's a pretty extensive commit message...