Lab 02 completed
Lab 02 exercise has been completed. Please, do not take this code seriously, there is literally duplicated code in the socket and stream creation/destruction.
I mean, it was just really a quick implementation, absolutely nothing fancy about the design. An alternative would be to implement some sort of "message termiantion" protocol, that would effectively replace EOF, but I'm not sure if we can modify the actual "protocol" implemented in the provided code.